Import of MSCHILLI/Log-Log4perl-0.33 from CPAN.
gitpan-cpan-distribution: Log-Log4perl
gitpan-cpan-version:      0.33
gitpan-cpan-path:         MSCHILLI/Log-Log4perl-0.33.tar.gz
gitpan-cpan-author:       MSCHILLI
gitpan-cpan-maturity:     released

+++ t/017Watch.t
@@ -8,6 +8,14 @@ use strict;
 use Log::Log4perl;
 use File::Spec;
 
+my $WORK_DIR = "tmp";
+if(-d "t") {
+    $WORK_DIR = File::Spec->catfile(qw(t tmp));
+}
+unless (-e "$WORK_DIR"){
+    mkdir("$WORK_DIR", 0755) || die "can't create $WORK_DIR ($!)";
+}
+
 my $testfile = File::Spec->catfile(qw(t tmp test17.log));
 unlink $testfile if (-e $testfile);
 
@@ -19,7 +27,7 @@ unlink $testconf if (-e $testconf);
 my $conf1 = <<EOL;
 log4j.category.animal.dog   = INFO, myAppender
 
-log4j.appender.myAppender          = Log::Dispatch::File
+log4j.appender.myAppender          = Log::Log4perl::Appender::File
 log4j.appender.myAppender.layout   = Log::Log4perl::Layout::SimpleLayout
 log4j.appender.myAppender.filename = $testfile
 log4j.appender.myAppender.mode     = append
@@ -46,7 +54,7 @@ sleep 3;
 my $conf2 = <<EOL;
 log4j.category.animal.dog   = DEBUG, myAppender
 
-log4j.appender.myAppender          = Log::Dispatch::File
+log4j.appender.myAppender          = Log::Log4perl::Appender::File
 log4j.appender.myAppender.layout = org.apache.log4j.PatternLayout
 log4j.appender.myAppender.layout.ConversionPattern=%-5p %c - %m%n
 
@@ -82,7 +90,7 @@ sleep 3;
 $conf2 = <<EOL;
 log4j.category.animal.dog   = INFO, myAppender
 
-log4j.appender.myAppender          = Log::Dispatch::File
+log4j.appender.myAppender          = Log::Log4perl::Appender::File
 log4j.appender.myAppender.layout   = Log::Log4perl::Layout::SimpleLayout
 log4j.appender.myAppender.filename = $testfile
 log4j.appender.myAppender.mode     = append
